refactor(hooks): move timeout followup state handling
